fix: :bug: move kudos admin panel from dashboard to kudos controller
