+++ app/views/layouts/_navbar.html.erb
@@ -36,10 +36,11 @@
       <% if admin_signed_in? %>
         <div class="navbar-item has-dropdown is-hoverable">
           <a class="navbar-link">
-            Admin options
+            Admin panel
           </a>
           <div class="navbar-dropdown is-right">
             <%= link_to 'Dashboard', admins_pages_dashboard_path, class: 'navbar-item' %>
+            <%= link_to 'Kudos', admins_kudos_path, class: 'navbar-item' %>
             <%= link_to 'Employees', admins_employees_path, class: 'navbar-item' %>
             <%= link_to 'Company Values', admins_company_values_path, class: 'navbar-item' %>
             <hr class="navbar-divider">
